pyproject.toml
[tool.poetry]
name = "ticker_monitor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]


[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
python-dotenv = "^0.18.0"
praw = "^7.4.0"
pytz = "^2021.1"
pymongo = {extras = ["srv"], version = "^3.12.0"}
tqdm = "^4.62.2"
pandas = "^1.3.2"
utility-soup = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
